requirements.txt
uvicorn[standard]
boto3==1.24.75
fastapi~=0.79.1
python-dotenv~=0.21.0
pydantic~=1.10.2
uuid~=1.30
botocore~=1.27.75
python-magic
python-multipart